Add Roles Abstraction and clean up Users module
