+++ src/OrchardCore/OrchardCore.Roles.Abstractions/IRoleRemovedEventHandler.cs
